(ccl_split)struct.lisp: Add a new SELECTION-SET-BY-COMMAND slot to the BUFFER structure.
cocoa-editor.lisp: clear the new SELECTION-SET-BY-COMMAND slot if
the selection's set by Cocoa; set if it the selection's set by Hemlock
command processing (and has non-zero length.)

bindings.lisp: new functions CURRENT-POINT-FOR-SELECTION-START,
CURRENT-POINT-FOR-SELECTION-END:  ensure that the Hemlock region
is active and return point; if the region was active but the
selection wasn't set by an editing command, ensure that point
is at the start (end) of the selection (exchanging the positions
of point and mark if necessary.)

package.lisp: export those function names.

command.lisp, lispmode.lisp: use these new functions for motion
commands that modify the selection.  As far as I can see, their
behavior is pretty close to TextEdit's (there -may- be cases where
TextEdit's notion of a "word" differs from Hemlock's, and there may
be different behavior regarding whether or not whitepace is included
in the selection in some cases.)

If this isn't broken and someone works on it: we currently pass
key events that have the command modifier set to Cocoa, and
command-shift-arrow extends the selection to the start/end of the
line/buffer.  We may need to either handle these events as Hemlock
commands (with a hyper or other modifer) or somehow ensure that
our notion of how the selection is anchored matches Cocoa's (which
seems to be maintained in some internal undocumented way.)

bindings.lisp: bind meta-leftarrow, meta-rightarrow.  (May still need
to setup bindings for other modified arrow-key events.)