(ccl_split)Fixed bug reported in the new Listener code:
Text copied down to the input from previous inputs or output no longer evaluates immediately, but is instead simply copied to input, and point is moved to the end of it. The complaint was that immediatel eval foreclosed the possibility of editing the previous input.

Still working on fixing scrolling bug and spurious insertion of newlines.